tests: split CMUXOpenCommandTests shared helpers into CMUXOpenCommandTestSupport
CMUXOpenCommandTests.swift is over the 900-line hard cap, so the +6 line
growth from the resolved-asset-URL assertions failed workflow-guard-tests.
Move the suite's shared process/socket/mock-server helpers into a new
CMUXOpenCommandTestSupport.swift extension file (wired into cmuxTests) and
ratchet the file's length budget down 2882 -> 2507.
